Extract current actions into SearchViewController
