+++ src/renderer/Core/Search/Helpers/getActions.ts
