+++ src/renderer/Core/Search/Helpers/index.ts
@@ -1,2 +1,5 @@
 export * from "./elementIsVisible";
+export * from "./getActions";
+export * from "./getNextSearchResultItemId";
+export * from "./getPreviousSearchResultItemId";
 export * from "./getSearchResultItemActionByKeyboardshortcut";
